Устранен warning C4018.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@5658 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/apps/mplayerc/PPageMouse.cpp
@@ -249,9 +249,9 @@ void CPPageMouse::OnDolabeleditList(NMHDR* pNMHDR, LRESULT* pResult)
 			const auto& ids = mouse_cmds->ids;
 			int nSel = -1;
 
-			for (int i = 0; i < ids.size(); i++) {
+			for (size_t i = 0; i < ids.size(); i++) {
 				if (id == ids[i]) {
-					nSel = i;
+					nSel = (int)i;
 					break;
 				}
 			}
